ec: drop public arch specific functions
ec_highlevel_func.c has code for SSE/AVX etc, and the headers have
prototypes for them as well.

We want the API to only contain arch-independent functions.

Break the highlevel x86 functions into their own file.
Move the remaining function into ec_base.c, delete ec_highlevel_func.c
Remove the x86-specific prototypes from the public headers.
Create a new, private header file with the prototypes.
